feat: Add Windows AD target module for on-prem domain controller
This commit adds a new module "windows_ad_target" to the main.tf file. The module is sourced from "app.terraform.io/tfo-apj-demos/target/boundary" with a version constraint of "~> 1.5".

The module is configured to create a Windows domain controller on-premises. It requires the following inputs:
- project_name: The name of the project
- hostname_prefix: The prefix for the hostname of the domain controller
- credential_store_token: The token for accessing the Vault credential store
- vault_address: The address of the Vault server

The module also requires a list of hosts, which are obtained from the "vm" module. Each host in the list has a hostname and an IP address.

Additionally, the module defines a service named "rdp" with type "tcp" and port 3389. It also specifies a credential path for accessing the Vault LDAP dynamic demo role.

This new module will enable the creation of Windows AD targets for the vsphere platform.