feat: Add vault_token resource for service account
Add a new resource `vault_token` to create a service account token with the `ldap_reader` and `revoke_lease` policies. This token has a period of 24 hours and is intended for service account purposes.

Refactor the `main.tf` file to include the new `vault_token` resource and update the version of the `windows_ad_target` module to 1.5.